requirements.txt
python-dotenv
openai
json
